+++ example/app/api/serializers.py
@@ -1,3 +1,5 @@
+import typing
+
 import saritasa_drf_tools.serializers
 
 from .. import models
@@ -13,7 +15,7 @@ class TestModelListSerializer(
         fields = (
             "id",
             "text_field",
-            "related_model_id",
+            "related_model",
         )
 
 
@@ -30,17 +32,36 @@ class RelatedTestModelSerializer(
 class TestModelDetailSerializer(
     saritasa_drf_tools.serializers.ModelBaseSerializer,
 ):
-    """List Serializer."""
-
-    related_model = RelatedTestModelSerializer(
-        read_only=True,
-    )
+    """Detail Serializer."""
 
     class Meta:
         model = models.TestModel
         fields = (
             "id",
             "text_field",
-            "related_model_id",
             "related_model",
         )
+        data_fields: typing.ClassVar = {
+            "related_model": (
+                "example.app.api.serializers.RelatedTestModelSerializer"
+            ),
+        }
+
+
+class RelatedTestModelWithManyRelatedSerializer(
+    saritasa_drf_tools.serializers.ModelBaseSerializer,
+):
+    """Serializer for related mode with list of test models."""
+
+    class Meta:
+        model = models.TestRelatedModel
+        fields = (
+            "id",
+            "text_field",
+            "test_models",
+        )
+        data_fields: typing.ClassVar = {
+            "test_models": (
+                "example.app.api.serializers.TestModelListSerializer"
+            ),
+        }