+++ saritasa_drf_tools/serializers/model_base.py
@@ -7,6 +7,7 @@ class ModelBaseSerializer(
     mixins.FieldMappingOverride,
     mixins.CleanValidationMixin,
     mixins.UserAndRequestFromContextMixin,
+    mixins.DataFieldsMixin,
     serializers.ModelSerializer,
 ):
     """Model Serializer with common logic."""