+++ CHANGELOG.md
@@ -5,6 +5,7 @@ We follow [Semantic Versions](https://semver.org/).
 ## Unreleased
 
 - Make `OrderingFilterBackend`/`SearchFilterBackend` show `action` with openapi warnings
+- Add `DataFieldsMixin` for serializers
 
 ## 0.1.0